requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# To update, run:
#
#    pip-compile requirements.in
#
anyio==4.2.0
    # via fastapi, starlette
bcrypt==4.1.2
    # via -r requirements.in
email-validator==2.1.1
    # via -r requirements.in
fastapi==0.110.0
    # via -r requirements.in
h11==0.14.0
    # via uvicorn
idna==3.6
    # via anyio, email-validator
jinja2==3.1.3
    # via -r requirements.in
markupsafe==2.1.5
    # via jinja2
pillow==10.2.0
    # via -r requirements.in
pydantic==2.6.4
    # via fastapi, pydantic-settings
pydantic-core==2.16.3
    # via pydantic
pydantic-settings==2.2.1
    # via -r requirements.in
python-dotenv==1.0.1
    # via -r requirements.in, pydantic-settings
python-jose==3.3.0
    # via -r requirements.in
python-multipart==0.0.9
    # via -r requirements.in, fastapi
redis==5.0.3
    # via -r requirements.in
requests==2.31.0
    # via -r requirements.in
rfc3986-validator==0.1.1
    # via python-jose
rsa==4.9
    # via python-jose
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.28
    # via -r requirements.in
starlette==0.36.3
    # via fastapi
ttkbootstrap==1.10.1
    # via -r requirements.in
typing-extensions==4.10.0
    # via pydantic, starlette, uvicorn
uvicorn==0.27.1
    # via -r requirements.in
asyncpg==0.29.0
    # via -r requirements.in
psycopg2-binary==2.9.11
    # PostgreSQL adapter for Python